requirements.txt
Django>=5.2.7
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9

